TFS_Update
git-svn-id: https://dotnetnuke.svn.codeplex.com/svn@61665 142e9c35-602f-4880-9f37-ec48a24daa6a

+++ Trunk/Library/UI/Skins/SkinThumbNailControl.cs
@@ -282,14 +282,14 @@ namespace DotNetNuke.UI.Skins
                     if (objImage.Height > objImage.Width)
                     {
 						//The height was larger, so scale the width 
-                        dblScale = intSize/objImage.Height;
+                        dblScale = (double)intSize / objImage.Height;
                         intHeight = intSize;
                         intWidth = Convert.ToInt32(objImage.Width*dblScale);
                     }
                     else
                     {
 						//The width was larger, so scale the height 
-                        dblScale = intSize/objImage.Width;
+                        dblScale = (double)intSize / objImage.Width;
                         intWidth = intSize;
                         intHeight = Convert.ToInt32(objImage.Height*dblScale);
                     }
